refactor to use move the test client and session fixture to its own file
